[#1599] Bring the customer definition back in too
