+++ apps/api/v2/models/customer/features/customer_safe_dump.rb
@@ -1,0 +1,17 @@
+# apps/api/v2/models/customer/features/customer_safe_dump.rb
+
+module V2
+  module Models
+    module Customer
+      module Features
+        module CustomerSafeDump
+
+          # TODO: Should the @safe_dump definition live here? We'd need a method
+          # to call I think  b/c just putting the @safe_dump class instance var
+          # here seems brittle all less formalized.
+
+        end
+      end
+    end
+  end
+end
